Handle incoming units in entity constructors taking entities as arguments.
When we create an entity that takes other potentially unitful entities
as parameters, we need to make sure that the resulting units are sound.
Principally, this means inspecting units when extracting coordinates/components
from points & vectors, and converting them to a commong unit.

TODO(andrew): we currently aggressively convert point coordinates to the
default units (meters); consider preserving incoming entity units when
possible.

Change-Id: I82fdc793fff2b28261b1bbcd63643d621efb8110