LIB3D-164: Revit Schema
In that CL I am defining a schema that would specify a
revit element. The Schema for revit elements is based
upon primitive geometry objects defined in psworker.json schema.

Schema for following element is defined:
revit-wall, revit-room, revit-referencePlane, revit-modelCurve,
revit-level, revit-floor, revit-oneLevelFamilyInstance,
revit-twoLevelFamilyInstance and revit-oneLevelHostedFamilyInstance.

I am also moving the revit.js library to flux-modelingjs
so that revit objects can be created consistently by the
blocks that are validated against the schema.

:not-maker-facing:

QA: This cannot be currently tested by QA. It would be testable when viewer
is updated to validate revit objects against this schema.

Change-Id: I352cca538158a51726404d7a2063a8d47f063466