LIB3D-748: ModelLine and Reference plane should accept flux Id.
In this CL I am making sure that ModelLine and Reference Plane elements are
able to receive FluxId. This is necessary for the merge functionality in the
plugin. Previously it was assumed that this elements would always be replaced which
is not true anymore.

:non-maker-facing:

Change-Id: I5e0fb29e99a02c32be6ed41c3d33a24a65f5434e
QA: This would be tested through the blocks using this module.