LIB3D-785: Revit profile should accept a polyline.
In this CL I am updating revit schema such that revit profile can have a
polyline as a valid entity.

:maker-facing:

QA: Create a revit floor and a wall polyline(s) as an input to the profile node
of the create block. Make sure that the corresponding object is created in
the revit model.

Change-Id: I61f78627d99e274ff575f78dcf8549f608a8a858