LIB3D-774. GetParameters block does not work gracefully.
GetParameters block should be more forgiving and not error out completely
when an invalid Revit Element is input. It should instead try to find the
requested parameter and return false if the parameter is not found.

:maker-facing:

QA: Make sure GetParameters block does not give error when an invalid Revit
Element is input. Instead it should just return false if the requested
parameter is not found.

Change-Id: Ide1f36aaf5229ef994c72f55cfa5a5c88bffdc33