LIB3D-811: fluxId should accept number as well as string.
I am changing the revit library such that it would accept the fluxId input
as a number as well as a string. It would automatically convert a number
input to string to confirm with the schema. This change itself is not
maker facing. Corresponding changes to revit blocks would be.

:non-maker-facing:

QA) None.

Change-Id: I7487866b5cc3f168f1646e2cc8bbfd64f69f2d05