LIB3D-834: Revit CreateFamilyInstance gives incorrect output.
In this CL I am fixing Revit Library such that methods that create family
instance object behave in the same manner as other revit blocks and give
appropriate error instead of wrong output object when an invalid input is
provided.

:maker-facing:

QA: Make sure that CreateOneLevelFamilyInstance, CreateTwoLevelFamilyInstance and
CreateOneLevelHostedFamilyInstance blocks gives out error when an invalid input is
given for the location input. Incorrect output object should not be created.

Change-Id: I807140475f7668289238114ef85d1a114a677f24