Move some units-handling stuff out of modelingjs and in to typecheck.js
This makes modeling.js free of direct references to various schemas.
It also reduces our flexibility in when schema's are required in. However,
as we bake more functionality into modelingjs (unit conversion, typechecking),
I believe that this is preferable to passing through the dynamic language
equivalents of #IFDEF's to enable or disable vast swaths of functionality
for relatively limited use-cases.

:not-maker-facing:

Change-Id: Iff5ccbbfd45d961e7927c7c075098193862e1ba4