JSON-schema based typechecking for entity constructors
This borrows heavily from https://cr.flux.io/#/c/12494/, although takes
a different approach in a few ways:
 * provide an explicit validation function, rather than decorating
function calls. This makes calling uniform between entity constructors
and places where we want to validate data types, but want the caller to
do something besides explode. The principle example of this is in "Burst"
blocks that decompose entities into the constituent parts; in these
cases, we want to detect that an input entity does not match it's spec,
and warn the user, but not throw an exception.
 * has a more limited set of functionality (Can't differentiate null and
undefined, can't match based on javascript classes, can't transform/coerce
values while typechecking, etc).
 * uses json-schema for all type descriptions

:not-maker-facing:

QA) None.

Change-Id: I2743be58a57f62dbc0505194b58c5ac85d97dd94