LIB3D-903: Adding "isSolid" property to mesh schema.
In this CL I am adding the "isSolid" property to the mesh schema. This
property would indicate if the geometry represented by the mesh should
be treated as a solid body or a sheet body. isSolid=true would mean that
the geometry would be treated as a solid body.

:not-maker-facing:

QA: None.

Change-Id: I94727dba93a316475a945adb20265a59765c1aaf