LIB3D-889 Get better about re-using entities during operation resolution.
https://vannevar.atlassian.net/browse/LIB3D-889

This also clears out the entity pseudoclass, and relies on existence of
the "primitive" key when determining whether or not a given object
is an entity.

There are still situations where entities might be encoded into queries
more than once -- this solution still relies on object comparison, not
a deep equality check. However, this is an improvement.

Change-Id: Ic29f3bb2a1503aded9da45d1789bc5232f101e11