Flatten Scene.
In this CL I am implementing Flatten operation for a scene.
The output of this flatten operation is an instance primitive that is
in an intermediate state such that entities are baked into
the instance, and attributes are merged but cumulative matrix is
not yet applied to the geometry.

:not-maker-facing:

QA: None.

Change-Id: I4fe04fb61e1f15b7f1e45e1874141a4b34f9cbbe