LIB3D-1000: Implement entity transform.
In this CL I am implementing transformations of Flux entities natively in JS library.
This will current transformation workflow in two ways:

1) The types won't be converted to Brep when a geometric entity is transformed.
2) We would not have to make a trip to PBW for transformations all flux entities
except Breps.

Some transformation operations such as reflection and projections that are not
supported on analytical geometry would return appropriate errors.

This is not maker facing until existing blocks are updated to use this library for
transformation instead of PBW.

:not-maker-facing:

Change-Id: Ib26e72f6840d6b0afbf6c826e5ffb123cdafbf8d
QA: None.