Specify parent in scene validator errors - LIB3D-1089 LIB3D-1061
npm version 0.8.3
When the scene validator reported an error that a child was not allowed
with a given parent, it was ambiguous what was wrong, since the type of
the parent was unknown. This extends the error message for clarity.
Also fix bug in prep where geometryList children with errors caused
an exception.

:hp: :viewport:

Code Review:
1. Check out related CLs.
2. make run-viewport
QA:
1. Render the liberty.json from JIRA.
2. The text of the error in the viewport should contain one message,
   not two.

Change-Id: Iff682fec86c891960ffd32ab5d700c8ed52441bf